enterprise-search/requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
chromadb>=0.4.0
python-dotenv>=1.0.0
rich>=13.0.0
pypdf>=3.0.0
tiktoken>=0.5.0
rank_bm25>=0.2.2
jieba>=0.42.1
sentence-transformers>=2.2.0
google-generativeai>=0.5.0
